Add useScrollAction hooks
