+++ src/types/pickerItemRef.ts
@@ -1,0 +1,3 @@
+import { RefObject } from "react";
+
+export type PickerItemRef = { [key: string]: RefObject<HTMLLIElement> };
